Rename and minor refactor
